Integration tests (#4)
* Written service integration tests for deleting a file in the FileService

* Added helper method for generating random string

* Added a project for common tests' setups

* Added tests hekoer method for generating a random string

* Written service integration tests for storing an image to a stream

* Code cleaning

* Code cleaning

* Code refactoring

* Written API integration tests for the CatBreedDetection controller

* Projects reorganization

* Fixed failing tests and a bug with the invalid model

* Added tests for the cat breeds classifier (WIP)

* Migrated the projects to NET7 and code refactoring

+++ src/CatBreedsDetector.Server/Controllers/CatBreedsController.cs
@@ -3,8 +3,6 @@
 namespace CatBreedsDetector.Web.Controllers
 {
     using System;
-    using System.IO;
-    using System.Reflection;
     using System.Threading.Tasks;
     using CatBreedsDetector.Classification.Interfaces;
     using CatBreedsDetector.Common;
@@ -18,6 +16,8 @@ namespace CatBreedsDetector.Web.Controllers
     [Route("api/[controller]")]
     public class CatBreedsController : ControllerBase
     {
+        private const string DirectoryPath = "C:\\Temp\\TestFiles";
+
         private readonly ICatBreedClassifier _catBreedClassifier;
 
         private readonly IFileService _fileService;
@@ -31,8 +31,10 @@ namespace CatBreedsDetector.Web.Controllers
         [HttpPost(nameof(DetectAsync))]
         public async Task<IActionResult> DetectAsync([FromForm] CatBreedDetectInputModel model, CancellationToken cancellationToken)
         {
-            var buildDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            var predictedImagesDirectoryPath = buildDir + $@"\{Constants.FilePath.PredictedImages}";
+            if (model is null || !this.ModelState.IsValid)
+                return this.BadRequest();
+
+            var predictedImagesDirectoryPath = DirectoryPath + $@"\{Constants.FilePath.PredictedImages}";
             var imagePath = predictedImagesDirectoryPath + $@"\{model.CatImage.FileName}";
 
             this._fileService.DeleteFilesInDirectory(predictedImagesDirectoryPath);
